+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hyperdx/cli
 
+## 0.1.1
+
+### Patch Changes
+
+- 458486b: feat: Add --apiVersion flag to configure API version (v1 or v2)
+
 ## 0.1.0
 
 ### Minor Changes